chore: move utils folder
