+++ internal/guest/handler/customer/create.go
@@ -4,8 +4,8 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/ua-academy-projects/share-bite/internal/guest/util/httpctx"
-	"github.com/ua-academy-projects/share-bite/internal/guest/util/request"
+	"github.com/ua-academy-projects/share-bite/internal/util/httpctx"
+	"github.com/ua-academy-projects/share-bite/internal/util/request"
 )
 
 func (h *handler) create(c *gin.Context) {
