+++ internal/guest/handler/post/create.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/ua-academy-projects/share-bite/internal/guest/entity"
-	"github.com/ua-academy-projects/share-bite/internal/guest/util/httpctx"
-	"github.com/ua-academy-projects/share-bite/internal/guest/util/request"
+	"github.com/ua-academy-projects/share-bite/internal/util/httpctx"
+	"github.com/ua-academy-projects/share-bite/internal/util/request"
 )
 
 type createRequest struct {
